test: move toggle button tests to dropdown suite
